Use dynamic memory allocation to load sketch "coredump" to memory

+++ Utils.cpp
@@ -282,7 +282,7 @@ int Utils::setNonblocking(int fd) {
     return fcntl(fd, F_SETFL, flags | O_NONBLOCK);
 }
 
-off_t Utils::fileSize(const char *filename) {
+size_t Utils::fileSize(const char *filename) {
 	struct stat st;
 	if (stat(filename, &st) == 0)
 		return st.st_size;
